backend/requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
certifi>=2023.7.22
flask-cors==4.0.1
gunicorn>=21.2.0